renaming file of useCurrentOrganization & adjusting front-end custody visibility on booking page
